# fix: LogLine.ToText not consistent