#### :question: What

The `.ToText` method of the `LogLine` doesn’t always give the same output, for what should be the same serialized logline.
If the `.Set` hasn’t been called (so no fields have been modified) the serialized CSV Logline will only contain the useful fields and stop after the last separator of non-nil fields, for example:
```
a,b,c,d
```

However, if Set is called, the serialized logline will contain as many separators as there are fields, for example:
```
a,b,c,d,,,,,,,,,, x 3000
```
#### :hammer: The fix

The PR slightly changes the `.Parse` and `.ToText` methods in order to know which is the last field and thus correctly stop. The fix seems to not impact the performance of both `.Parser` and `.ToText` methods. Here are the results:

 ```
name                        old time/op    new time/op    delta
LogLineParse-8                2.01µs ± 0%    1.42µs ± 0%   ~     (p=1.000 n=1+1)
LogLineToText/from_set-8      2.02µs ± 0%    3.04µs ± 0%   ~     (p=1.000 n=1+1)
LogLineToText/from_parse-8     475ns ± 0%     477ns ± 0%   ~     (p=1.000 n=1+1)
LogLineCopy/set=0-8           7.98µs ± 0%    9.29µs ± 0%   ~     (p=1.000 n=1+1)
LogLineCopy/set=1-8           41.7µs ± 0%    15.3µs ± 0%   ~     (p=1.000 n=1+1)
LogLineCopy/set=10-8          30.7µs ± 0%    15.4µs ± 0%   ~     (p=1.000 n=1+1)
LogLineCopy/set=20-8          32.2µs ± 0%    22.6µs ± 0%   ~     (p=1.000 n=1+1)

name                        old alloc/op   new alloc/op   delta
LogLineParse-8                 0.00B          0.00B        ~     (all equal)
LogLineToText/from_set-8      1.54kB ± 0%    1.54kB ± 0%   ~     (all equal)
LogLineToText/from_parse-8    1.28kB ± 0%    1.28kB ± 0%   ~     (all equal)
LogLineCopy/set=0-8           24.5kB ± 0%    24.5kB ± 0%   ~     (all equal)
LogLineCopy/set=1-8           34.5kB ± 0%    25.0kB ± 0%   ~     (p=1.000 n=1+1)
LogLineCopy/set=10-8          34.5kB ± 0%    25.0kB ± 0%   ~     (p=1.000 n=1+1)
LogLineCopy/set=20-8          34.5kB ± 0%    29.1kB ± 0%   ~     (p=1.000 n=1+1)

name                        old allocs/op  new allocs/op  delta
LogLineParse-8                  0.00           0.00        ~     (all equal)
LogLineToText/from_set-8        2.00 ± 0%      2.00 ± 0%   ~     (all equal)
LogLineToText/from_parse-8      1.00 ± 0%      1.00 ± 0%   ~     (all equal)
LogLineCopy/set=0-8             3.00 ± 0%      3.00 ± 0%   ~     (all equal)
LogLineCopy/set=1-8             5.00 ± 0%      3.00 ± 0%   ~     (p=1.000 n=1+1)
LogLineCopy/set=10-8            5.00 ± 0%      3.00 ± 0%   ~     (p=1.000 n=1+1)
LogLineCopy/set=20-8            5.00 ± 0%      4.00 ± 0%   ~     (p=1.000 n=1+1)
```

There is also some new units test and a fix on the `TestLogLineParse_separators` tests which does not check the correct thing. 

#### :white_check_mark: Checklists

_This section contains a list of checklists for common uses, please delete the checklists that are useless for your current use case (or add another checklist if your use case isn't covered yet)._

- [x] Is there unit/integration test coverage for all new and/or changed functionality added in this PR?
- [x] Have the changes in this PR been functionally tested?
- [x] Has `make gofmt-write` been run on the code?
- [x] Has `make govet` been run on the code? Has the code been fixed accordingly to the output?
- [ ] Have the changes been added to the [CHANGELOG.md](./CHANGELOG.md) file?
- [x] Have the steps in [CONTRIBUTING.md](./CONTRIBUTING.md) been followed to update a Go module?