@arl 2021-07-24
@lucarin91 
I haven't had time to look at the code, i was just curious at the benchmarks, they seem promising but the last column, delta, doesn't show any difference since benchstat can't ensure results is statistically significant.

In other words you need to run the benchmarks with the `-count`. A count of 10 or 15 is generally enough, depending on the cases, and on the noise generated by your machine. 

This short talk is a good watch on the subject, and some tips to run Go benchmarks effectively
https://youtu.be/jiXnzkAzy30