@lucarin91 2021-08-11
### :up: benchmark updates
Improve and run a new benchmark considering different sizes of the parsed buffer and set values. The following benchmark confirms the previous results: the PR improves the overall performance of both `Parse` and `ToText` functions especially on loglines with a small number of fields.
For what concerns allocations the PR fixes an allocation bug that on loglines with a small number of fields end up allocating 5 times more space.

 ```
name                                            old time/op    new time/op    delta
LogLineParse/len=1-8                               721ns ± 2%       8ns ± 2%  -98.95%  (p=0.000 n=18+16)
LogLineParse/len=50-8                             1.07µs ± 3%    0.31µs ± 2%  -70.86%  (p=0.000 n=18+19)
LogLineParse/len=500-8                            3.65µs ± 1%    3.02µs ± 1%  -17.39%  (p=0.000 n=18+19)
LogLineParse/len=1000-8                           6.52µs ± 1%    6.03µs ± 2%   -7.44%  (p=0.000 n=18+19)
LogLineParse/len=2000-8                           12.3µs ± 1%    12.1µs ± 2%   -1.86%  (p=0.000 n=19+19)
LogLineParse/len=3000-8                           18.2µs ± 2%    18.2µs ± 3%     ~     (p=0.994 n=18+19)
LogLineToTextFromSet/set=1-8                       905ns ± 8%     837ns ± 9%   -7.52%  (p=0.000 n=20+20)
LogLineToTextFromSet/set=50-8                     1.67µs ± 3%    1.27µs ± 1%  -23.72%  (p=0.000 n=19+16)
LogLineToTextFromSet/set=100-8                    2.47µs ± 6%    1.74µs ± 2%  -29.66%  (p=0.000 n=19+18)
LogLineToTextFromSet/set=254-8                    4.14µs ± 1%    3.08µs ± 1%  -25.52%  (p=0.000 n=19+17)
LogLineToTextFromParse/parse=1-8                  32.9ns ± 1%    33.1ns ± 1%   +0.46%  (p=0.001 n=19+19)
LogLineToTextFromParse/parse=50-8                  188ns ± 3%     189ns ± 5%     ~     (p=0.465 n=20+20)
LogLineToTextFromParse/parse=500-8                1.39µs ± 2%    1.39µs ± 2%     ~     (p=0.542 n=19+18)
LogLineToTextFromParse/parse=1000-8               2.74µs ± 1%    2.75µs ± 2%     ~     (p=0.121 n=19+19)
LogLineToTextFromParse/parse=2000-8               5.52µs ± 1%    5.55µs ± 1%   +0.58%  (p=0.019 n=20+20)
LogLineToTextFromParse/parse=3000-8               7.22µs ± 1%    7.25µs ± 2%     ~     (p=0.153 n=19+18)
LogLineToTextFromParseSet/parse=100,set=50-8      20.2µs ± 2%     2.8µs ± 2%  -86.26%  (p=0.000 n=19+19)
LogLineToTextFromParseSet/parse=100,set=100-8     19.5µs ± 1%     2.8µs ± 0%  -85.67%  (p=0.000 n=19+17)
LogLineToTextFromParseSet/parse=100,set=254-8     19.5µs ± 4%     3.9µs ± 1%  -79.75%  (p=0.000 n=19+19)
LogLineToTextFromParseSet/parse=1000,set=50-8     25.7µs ± 2%    11.2µs ± 2%  -56.54%  (p=0.000 n=17+17)
LogLineToTextFromParseSet/parse=1000,set=100-8    26.6µs ± 3%    11.1µs ± 2%  -58.19%  (p=0.000 n=18+19)
LogLineToTextFromParseSet/parse=1000,set=254-8    21.3µs ± 2%    11.2µs ± 1%  -47.55%  (p=0.000 n=20+19)
LogLineToTextFromParseSet/parse=3000,set=50-8     32.9µs ± 1%    30.6µs ± 0%   -6.78%  (p=0.000 n=18+18)
LogLineToTextFromParseSet/parse=3000,set=100-8    32.9µs ± 3%    30.7µs ± 3%   -6.62%  (p=0.000 n=18+19)
LogLineToTextFromParseSet/parse=3000,set=254-8    35.1µs ±21%    30.3µs ± 1%  -13.48%  (p=0.000 n=20+18)

name                                            old alloc/op   new alloc/op   delta
LogLineParse/len=1-8                               0.00B          0.00B          ~     (all equal)
LogLineParse/len=50-8                              0.00B          0.00B          ~     (all equal)
LogLineParse/len=500-8                             0.00B          0.00B          ~     (all equal)
LogLineParse/len=1000-8                            0.00B          0.00B          ~     (all equal)
LogLineParse/len=2000-8                            0.00B          0.00B          ~     (all equal)
LogLineParse/len=3000-8                            0.00B          0.00B          ~     (all equal)
LogLineToTextFromSet/set=1-8                       40.0B ± 0%     16.0B ± 0%  -60.00%  (p=0.000 n=20+20)
LogLineToTextFromSet/set=50-8                     1.54kB ± 0%    0.58kB ± 0%  -62.50%  (p=0.000 n=20+20)
LogLineToTextFromSet/set=100-8                    3.07kB ± 0%    1.15kB ± 0%  -62.50%  (p=0.000 n=20+20)
LogLineToTextFromSet/set=254-8                    5.89kB ± 0%    3.07kB ± 0%  -47.83%  (p=0.000 n=20+20)
LogLineToTextFromParse/parse=1-8                   16.0B ± 0%     16.0B ± 0%     ~     (all equal)
LogLineToTextFromParse/parse=50-8                   576B ± 0%      576B ± 0%     ~     (all equal)
LogLineToTextFromParse/parse=500-8                6.14kB ± 0%    6.14kB ± 0%     ~     (all equal)
LogLineToTextFromParse/parse=1000-8               12.3kB ± 0%    12.3kB ± 0%     ~     (all equal)
LogLineToTextFromParse/parse=2000-8               24.6kB ± 0%    24.6kB ± 0%     ~     (all equal)
LogLineToTextFromParse/parse=3000-8               41.0kB ± 0%    41.0kB ± 0%     ~     (all equal)
LogLineToTextFromParseSet/parse=100,set=50-8      14.6kB ± 0%     1.8kB ± 0%  -87.72%  (p=0.000 n=20+20)
LogLineToTextFromParseSet/parse=100,set=100-8     13.1kB ± 0%     2.0kB ± 0%  -84.31%  (p=0.000 n=20+20)
LogLineToTextFromParseSet/parse=100,set=254-8     12.7kB ± 0%     3.5kB ± 0%  -72.73%  (p=0.000 n=20+20)
LogLineToTextFromParseSet/parse=1000,set=50-8     26.6kB ± 0%    12.3kB ± 0%  -53.85%  (p=0.000 n=20+20)
LogLineToTextFromParseSet/parse=1000,set=100-8    28.7kB ± 0%    12.3kB ± 0%  -57.14%  (p=0.000 n=20+20)
LogLineToTextFromParseSet/parse=1000,set=254-8    13.6kB ± 0%    13.6kB ± 0%     ~     (all equal)
LogLineToTextFromParseSet/parse=3000,set=50-8     41.0kB ± 0%    41.0kB ± 0%     ~     (all equal)
LogLineToTextFromParseSet/parse=3000,set=100-8    41.0kB ± 0%    41.0kB ± 0%     ~     (all equal)
LogLineToTextFromParseSet/parse=3000,set=254-8    41.0kB ± 0%    41.0kB ± 0%     ~     (all equal)

name                                            old allocs/op  new allocs/op  delta
LogLineParse/len=1-8                                0.00           0.00          ~     (all equal)
LogLineParse/len=50-8                               0.00           0.00          ~     (all equal)
LogLineParse/len=500-8                              0.00           0.00          ~     (all equal)
LogLineParse/len=1000-8                             0.00           0.00          ~     (all equal)
LogLineParse/len=2000-8                             0.00           0.00          ~     (all equal)
LogLineParse/len=3000-8                             0.00           0.00          ~     (all equal)
LogLineToTextFromSet/set=1-8                        2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=20+20)
LogLineToTextFromSet/set=50-8                       2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=20+20)
LogLineToTextFromSet/set=100-8                      2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=20+20)
LogLineToTextFromSet/set=254-8                      2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=20+20)
LogLineToTextFromParse/parse=1-8                    1.00 ± 0%      1.00 ± 0%     ~     (all equal)
LogLineToTextFromParse/parse=50-8                   1.00 ± 0%      1.00 ± 0%     ~     (all equal)
LogLineToTextFromParse/parse=500-8                  1.00 ± 0%      1.00 ± 0%     ~     (all equal)
LogLineToTextFromParse/parse=1000-8                 1.00 ± 0%      1.00 ± 0%     ~     (all equal)
LogLineToTextFromParse/parse=2000-8                 1.00 ± 0%      1.00 ± 0%     ~     (all equal)
LogLineToTextFromParse/parse=3000-8                 1.00 ± 0%      1.00 ± 0%     ~     (all equal)
LogLineToTextFromParseSet/parse=100,set=50-8        5.00 ± 0%      1.00 ± 0%  -80.00%  (p=0.000 n=20+20)
LogLineToTextFromParseSet/parse=100,set=100-8       4.00 ± 0%      1.00 ± 0%  -75.00%  (p=0.000 n=20+20)
LogLineToTextFromParseSet/parse=100,set=254-8       3.00 ± 0%      1.00 ± 0%  -66.67%  (p=0.000 n=20+20)
LogLineToTextFromParseSet/parse=1000,set=50-8       2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=20+20)
LogLineToTextFromParseSet/parse=1000,set=100-8      2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=20+20)
LogLineToTextFromParseSet/parse=1000,set=254-8      1.00 ± 0%      1.00 ± 0%     ~     (all equal)
LogLineToTextFromParseSet/parse=3000,set=50-8       1.00 ± 0%      1.00 ± 0%     ~     (all equal)
LogLineToTextFromParseSet/parse=3000,set=100-8      1.00 ± 0%      1.00 ± 0%     ~     (all equal)
LogLineToTextFromParseSet/parse=3000,set=254-8      1.00 ± 0%      1.00 ± 0%     ~     (all equal)
```