@lucarin91 2021-08-12
### :up:  New benchmarks results:

For unknown reasons after the last fix the `ToText` function drastically reduces its performance in the case of a low number of set operations. The allocations remain unchanged so I omitted them.

```
name                                            old time/op    new time/op    delta
LogLineParse/len=1-8                               721ns ± 2%       8ns ±13%  -98.88%  (p=0.000 n=18+20)
LogLineParse/len=50-8                             1.07µs ± 3%    0.31µs ± 1%  -70.98%  (p=0.000 n=18+20)
LogLineParse/len=500-8                            3.65µs ± 1%    3.01µs ± 1%  -17.42%  (p=0.000 n=18+20)
LogLineParse/len=1000-8                           6.52µs ± 1%    6.04µs ± 1%   -7.37%  (p=0.000 n=18+16)
LogLineParse/len=2000-8                           12.3µs ± 1%    12.1µs ± 2%   -1.97%  (p=0.000 n=19+19)
LogLineParse/len=3000-8                           18.2µs ± 2%    18.2µs ± 2%     ~     (p=0.654 n=18+20)
LogLineToTextFromSet/set=1-8                       905ns ± 8%    1642ns ± 6%  +81.43%  (p=0.000 n=20+20)
LogLineToTextFromSet/set=50-8                     1.67µs ± 3%    2.04µs ± 1%  +22.22%  (p=0.000 n=19+16)
LogLineToTextFromSet/set=100-8                    2.47µs ± 6%    2.48µs ± 1%   +0.35%  (p=0.027 n=19+19)
LogLineToTextFromSet/set=254-8                    4.14µs ± 1%    3.79µs ± 2%   -8.47%  (p=0.000 n=19+20)
LogLineToTextFromParse/parse=1-8                  32.9ns ± 1%    33.2ns ± 1%   +0.97%  (p=0.000 n=19+18)
LogLineToTextFromParse/parse=50-8                  188ns ± 3%     192ns ± 7%   +1.99%  (p=0.007 n=20+18)
LogLineToTextFromParse/parse=500-8                1.39µs ± 2%    1.38µs ± 1%     ~     (p=0.475 n=19+16)
LogLineToTextFromParse/parse=1000-8               2.74µs ± 1%    2.75µs ± 1%     ~     (p=0.079 n=19+18)
LogLineToTextFromParse/parse=2000-8               5.52µs ± 1%    5.53µs ± 1%     ~     (p=0.512 n=20+17)
LogLineToTextFromParse/parse=3000-8               7.22µs ± 1%    7.26µs ± 2%   +0.45%  (p=0.029 n=19+17)
LogLineToTextFromParseSet/parse=100,set=50-8      20.2µs ± 2%     3.6µs ± 2%  -82.43%  (p=0.000 n=19+19)
LogLineToTextFromParseSet/parse=100,set=100-8     19.5µs ± 1%     3.6µs ± 3%  -81.65%  (p=0.000 n=19+20)
LogLineToTextFromParseSet/parse=100,set=254-8     19.5µs ± 4%     4.6µs ± 1%  -76.13%  (p=0.000 n=19+18)
LogLineToTextFromParseSet/parse=1000,set=50-8     25.7µs ± 2%    12.0µs ± 4%  -53.25%  (p=0.000 n=17+18)
LogLineToTextFromParseSet/parse=1000,set=100-8    26.6µs ± 3%    11.9µs ± 2%  -55.08%  (p=0.000 n=18+20)
LogLineToTextFromParseSet/parse=1000,set=254-8    21.3µs ± 2%    13.3µs ±19%  -37.48%  (p=0.000 n=20+20)
LogLineToTextFromParseSet/parse=3000,set=50-8     32.9µs ± 1%    41.0µs ± 1%  +24.67%  (p=0.000 n=18+17)
LogLineToTextFromParseSet/parse=3000,set=100-8    32.9µs ± 3%    41.0µs ± 4%  +24.59%  (p=0.000 n=18+20)
LogLineToTextFromParseSet/parse=3000,set=254-8    35.1µs ±21%    40.5µs ± 4%  +15.35%  (p=0.000 n=20+20)
```