@lucarin91 2021-08-13
#### :up: New Benchmarks results

After the last @arl fix the benchmark goes back to the old figures. In the following the time results for the complete benchmark go here [bench-last.txt](https://github.com/AdRoll/baker/files/6982608/bench-last.txt)

```
name                                            old time/op    new time/op    delta
LogLineParse/len=1-8                               731ns ± 2%       9ns ±19%  -98.84%  (p=0.000 n=19+20)
LogLineParse/len=50-8                             1.08µs ± 2%    0.31µs ± 2%  -70.99%  (p=0.000 n=18+19)
LogLineParse/len=500-8                            3.65µs ± 1%    3.03µs ± 1%  -17.05%  (p=0.000 n=19+19)
LogLineParse/len=1000-8                           6.57µs ± 2%    6.06µs ± 2%   -7.82%  (p=0.000 n=18+17)
LogLineParse/len=2000-8                           12.3µs ± 2%    12.1µs ± 2%   -1.59%  (p=0.000 n=19+20)
LogLineParse/len=3000-8                           18.1µs ± 2%    18.2µs ± 2%     ~     (p=0.073 n=19+19)
LogLineToTextFromSet/set=1-8                       911ns ±13%     837ns ± 7%   -8.12%  (p=0.000 n=20+20)
LogLineToTextFromSet/set=5-8                       952ns ± 3%     848ns ± 1%  -10.99%  (p=0.000 n=20+18)
LogLineToTextFromSet/set=50-8                     1.69µs ± 2%    1.29µs ± 1%  -23.68%  (p=0.000 n=19+19)
LogLineToTextFromSet/set=100-8                    2.48µs ± 3%    1.76µs ± 3%  -28.96%  (p=0.000 n=19+18)
LogLineToTextFromSet/set=254-8                    4.20µs ± 2%    3.10µs ± 1%  -26.00%  (p=0.000 n=18+18)
LogLineToTextFromParse/parse=1-8                  32.4ns ± 2%    32.8ns ± 1%   +1.26%  (p=0.000 n=18+18)
LogLineToTextFromParse/parse=50-8                  247ns ±26%     191ns ± 5%  -22.92%  (p=0.000 n=20+20)
LogLineToTextFromParse/parse=500-8                1.80µs ± 5%    1.40µs ± 3%  -22.39%  (p=0.000 n=17+19)
LogLineToTextFromParse/parse=1000-8               3.63µs ± 5%    2.76µs ± 1%  -23.97%  (p=0.000 n=20+19)
LogLineToTextFromParse/parse=2000-8               7.29µs ± 8%    5.55µs ± 2%  -23.88%  (p=0.000 n=19+20)
LogLineToTextFromParse/parse=3000-8               9.23µs ± 7%    8.33µs ±15%     ~     (p=0.087 n=20+20)
LogLineToTextFromParseSet/parse=100,set=5-8       26.8µs ± 4%     3.4µs ± 5%  -87.44%  (p=0.000 n=18+18)
LogLineToTextFromParseSet/parse=100,set=50-8      25.9µs ± 2%     3.5µs ± 3%  -86.34%  (p=0.000 n=19+19)
LogLineToTextFromParseSet/parse=100,set=100-8     25.0µs ± 2%     3.6µs ± 2%  -85.64%  (p=0.000 n=19+19)
LogLineToTextFromParseSet/parse=100,set=254-8     24.8µs ± 1%     5.1µs ± 1%  -79.49%  (p=0.000 n=18+19)
LogLineToTextFromParseSet/parse=1000,set=5-8      32.5µs ± 5%    14.7µs ± 7%  -54.76%  (p=0.000 n=19+18)
LogLineToTextFromParseSet/parse=1000,set=50-8     32.3µs ± 2%    14.5µs ± 2%  -55.04%  (p=0.000 n=19+20)
LogLineToTextFromParseSet/parse=1000,set=100-8    33.3µs ± 1%    14.4µs ± 1%  -56.76%  (p=0.000 n=19+19)
LogLineToTextFromParseSet/parse=1000,set=254-8    26.8µs ± 3%    14.5µs ± 3%  -45.81%  (p=0.000 n=18+19)
LogLineToTextFromParseSet/parse=3000,set=5-8      41.9µs ± 1%    40.2µs ± 6%   -4.22%  (p=0.000 n=19+19)
LogLineToTextFromParseSet/parse=3000,set=50-8     41.8µs ± 1%    39.8µs ± 2%   -4.81%  (p=0.000 n=18+19)
LogLineToTextFromParseSet/parse=3000,set=100-8    42.1µs ± 3%    39.8µs ± 3%   -5.43%  (p=0.000 n=18+20)
LogLineToTextFromParseSet/parse=3000,set=254-8    41.6µs ± 1%    39.5µs ± 4%   -5.00%  (p=0.000 n=19+18)
```